backend/requirements.txt
# Backend FastAPI - Requisitos

# Framework Web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Banco de Dados
sqlalchemy==1.4.51
alembic==1.12.1

# Validação
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.1
requests==2.31.0

# Utilitários
python-dotenv==1.0.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
openpyxl==3.1.2

# Já instalados no projeto principal
# numpy
# xlsxwriter
